Allow user to pass arguments to bootstrap.sh and build_all.sh for CMAKE
Also add checks so we don't keep downloading unneeded files

+++ CCAKE/bootstrap.sh
@@ -6,20 +6,31 @@ EoSdir=EoS/Houston
 
 #-------------------------------------------------------------------------------
 # set up example ICCING initial conditions
-mkdir -p ${ICdir}
+# if IC file does not exist
+if [ ! -f ${ICdir}/ICCING_example.dat ]
+then
+  mkdir -p ${ICdir}
 
-ICCINGexample="https://zenodo.org/record/6829000/files/ICCING_example.dat?download=1"
-wget -O ${ICdir}/ICCING_example.dat ${ICCINGexample}
+  ICCINGexample="https://zenodo.org/record/6829000/files/ICCING_example.dat?download=1"
+  wget -O ${ICdir}/ICCING_example.dat ${ICCINGexample}
+fi
 
 #-------------------------------------------------------------------------------
-# set up lattice-based EoS table (text and HDF format)
-mkdir -p ${EoSdir}
-
-EoSTXTTable="https://zenodo.org/record/6829115/files/thermo.dat?download=1"
-EoSHDFTable="https://zenodo.org/record/6829115/files/thermo.h5?download=1"
-wget -O ${EoSdir}/thermo.dat ${EoSTXTTable}
-wget -O ${EoSdir}/thermo.h5 ${EoSHDFTable}
-
+# set up lattice-based EoS tables
+# if TXT file does not exist
+if [ ! -f ${EoSdir}/thermo.dat ]
+then
+  mkdir -p ${EoSdir}
+  EoSTXTTable="https://zenodo.org/record/6829115/files/thermo.dat?download=1"
+  wget -O ${EoSdir}/thermo.dat ${EoSTXTTable}
+fi
+# if HDF file does not exist
+if [ ! -f ${EoSdir}/thermo.h5 ]
+then
+  mkdir -p ${EoSdir}
+  EoSHDFTable="https://zenodo.org/record/6829115/files/thermo.h5?download=1"
+  wget -O ${EoSdir}/thermo.h5 ${EoSHDFTable}
+fi
 
 #-------------------------------------------------------------------------------
 # make all shell scripts executable
@@ -28,4 +39,4 @@ chmod +x *.sh
 
 #-------------------------------------------------------------------------------
 # compile with CMake
-./build_all.sh
+./build_all.sh "$@"
